# swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: account-controller
    description: Account Controller
  - name: transaction-controller
    description: Transaction Controller
paths:
  /accounts:
    post:
      tags:
        - account-controller
      summary: save
      operationId: saveUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/AccountRequestDTO'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/AccountResponseDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/accounts/accounts/{id}':
    get:
      tags:
        - account-controller
      summary: buscar por id
      operationId: findByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AccountResponseDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /transactions:
    post:
      tags:
        - transaction-controller
      summary: save
      operationId: saveUsingPOST_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: dto
          description: dto
          required: true
          schema:
            $ref: '#/definitions/TransactionRequestDTO'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/TransactionResponseDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  AccountRequestDTO:
    type: object
    properties:
      documentNumber:
        type: string
    title: AccountRequestDTO
  AccountResponseDTO:
    type: object
    properties:
      accountId:
        type: integer
        format: int64
      documentNumber:
        type: integer
        format: int64
    title: AccountResponseDTO
  TransactionRequestDTO:
    type: object
    properties:
      accountId:
        type: integer
        format: int64
      amount:
        type: number
      operationTypeId:
        type: integer
        format: int64
    title: TransactionRequestDTO
  TransactionResponseDTO:
    type: object
    properties:
      accountId:
        type: integer
        format: int64
      amount:
        type: number
      eventDate:
        type: string
      operationTypeId:
        type: integer
        format: int64
    title: TransactionResponseDTO